Revert "Some refactoring the make the name of this use_case more true."
