--- src/main/java/interface_adapter/assisted_mode/AssistedModeColourConstants.java
@@ -1,8 +1,0 @@
-package interface_adapter.assisted_mode;
-
-import java.awt.Color;
-
-public class AssistedModeColourConstants {
-
-    public static final Color TABLECOLOUR = new Color(42, 99, 39);
-}
