+++ src/main/java/interface_adapter/assisted_mode/AssistedModeController.java
@@ -1,5 +1,4 @@
 package interface_adapter.assisted_mode;
 
 public class AssistedModeController {
-
 }
