+++ src/main/java/interface_adapter/assisted_mode/AssistedModeViewModel.java
@@ -1,11 +1,5 @@
 package interface_adapter.assisted_mode;
 
-import interface_adapter.ViewModel;
+public class AssistedModeViewModel {
 
-public class AssistedModeViewModel extends ViewModel<AssistedModeState> {
-
-    public AssistedModeViewModel() {
-        super("assisted_mode");
-        setState(new AssistedModeState());
-    }
 }
