--- src/main/java/interface_adapter/assisted_mode/hit/AssistedModeHitController.java
@@ -1,16 +1,0 @@
-package interface_adapter.assisted_mode.hit;
-
-import use_case.assisted_mode.hit.AssistedHitInputDataBoundary;
-
-public class AssistedModeHitController {
-
-    private final AssistedHitInputDataBoundary assistedHitInputDataBoundary;
-
-    public AssistedModeHitController(AssistedHitInputDataBoundary assistedHitInputDataBoundary) {
-        this.assistedHitInputDataBoundary = assistedHitInputDataBoundary;
-    }
-
-    public void execute() {
-        assistedHitInputDataBoundary.execute();
-    }
-}
