--- src/main/java/interface_adapter/assisted_mode/setup/AssistedModeSetUpPresenter.java
@@ -1,4 +1,0 @@
-package interface_adapter.assisted_mode.setup;
-
-public class AssistedModeSetUpPresenter {
-}
