+++ src/main/java/interface_adapter/freeplay/stand/FreePlayStandViewModel.java
@@ -1,0 +1,11 @@
+package interface_adapter.freeplay.stand;
+
+import interface_adapter.ViewModel;
+
+public class FreePlayStandViewModel extends ViewModel<FreePlayStandState> {
+
+    public FreePlayStandViewModel() {
+        super("stand");
+        setState(new FreePlayStandState());
+    }
+}
