+++ src/main/java/interface_adapter/probability/ProbabilityColourConstants.java
@@ -1,0 +1,10 @@
+package interface_adapter.probability;
+
+import java.awt.Color;
+
+public class ProbabilityColourConstants {
+
+    public static final Color TABLECOLOUR = new Color(42, 99, 39);
+    public static final Color PROGRESSCOLOUR = new Color(237, 202, 1);
+    public static final Color NONPROGRESSEDCOLOUR = new Color(182, 179, 179);
+}
