+++ src/main/java/interface_adapter/probability/hit/ProbabilityHitState.java
@@ -1,0 +1,16 @@
+package interface_adapter.probability.hit;
+
+/**
+ * The state for the ProbabiltyHit portion of ProbabiltyView.
+ */
+public class ProbabilityHitState {
+    private int hitProbability;
+
+    public int getStandProbability() {
+        return hitProbability;
+    }
+
+    public void setStandProbability(int standProbability) {
+        this.hitProbability = standProbability;
+    }
+}
