+++ src/main/java/interface_adapter/probability/hit/ProbabilityHitViewModel.java
@@ -1,0 +1,11 @@
+package interface_adapter.probability.hit;
+
+import interface_adapter.ViewModel;
+
+public class ProbabilityHitViewModel extends ViewModel<ProbabilityHitState> {
+
+    public ProbabilityHitViewModel() {
+        super("ProbabilityHit");
+        this.setState(new ProbabilityHitState());
+    }
+}
