+++ src/main/java/interface_adapter/probability/stand/ProbabilityStandState.java
@@ -1,0 +1,16 @@
+package interface_adapter.probability.stand;
+
+/**
+ * The state for the ProbabilityStand portion of the Probability View Model
+ */
+public class ProbabilityStandState {
+    private int standProbability;
+
+    public int getStandProbability() {
+        return standProbability;
+    }
+
+    public void setStandProbability(int standProbability) {
+        this.standProbability = standProbability;
+    }
+}
