+++ src/main/java/interface_adapter/probability/stand/ProbabilityStandViewModel.java
@@ -1,0 +1,11 @@
+package interface_adapter.probability.stand;
+
+import interface_adapter.ViewModel;
+
+public class ProbabilityStandViewModel extends ViewModel<ProbabilityStandState> {
+
+    public ProbabilityStandViewModel() {
+        super("ProbabilityStand");
+        this.setState(new ProbabilityStandState());
+    }
+}
