--- src/main/java/use_case/assisted_mode/hit/AssistedHitInputDataBoundary.java
@@ -1,6 +1,0 @@
-package use_case.assisted_mode.hit;
-
-public interface AssistedHitInputDataBoundary {
-
-    void execute();
-}
