--- src/main/java/use_case/dealer_screen/DealerScreenInputBoundary.java
@@ -1,12 +1,0 @@
-package use_case.dealer_screen;
-
-/**
- * The Input Boundary for the FreePlayStand Use case.
- */
-public interface DealerScreenInputBoundary {
-
-    /**
-     * Execute the FreePlayStand Use case.
-     */
-    void execute();
-}
