--- src/main/java/use_case/dealer_screen/DealerScreenOutputDataBoundary.java
@@ -1,9 +1,0 @@
-package use_case.dealer_screen;
-
-public interface DealerScreenOutputDataBoundary {
-
-    /**
-     * Prepares for the Exit View.
-     */
-    void prepareStandView(DealerScreenOutputData outputData);
-}
