+++ src/main/java/use_case/freeplay/stand/FreePlayStandInputBoundary.java
@@ -1,0 +1,12 @@
+package use_case.freeplay.stand;
+
+/**
+ * The Input Boundary for the FreePlayStand Use case.
+ */
+public interface FreePlayStandInputBoundary {
+
+    /**
+     * Execute the FreePlayStand Use case.
+     */
+    void execute();
+}
