+++ src/main/java/use_case/freeplay/stand/FreePlayStandOutputDataBoundary.java
@@ -1,0 +1,9 @@
+package use_case.freeplay.stand;
+
+public interface FreePlayStandOutputDataBoundary {
+
+    /**
+     * Prepares for the Exit View.
+     */
+    void prepareStandView(FreePlayStandOutputData outputData);
+}
