+++ src/main/java/use_case/probability/hit/ProbabilityHitInputBoundary.java
@@ -1,0 +1,12 @@
+package use_case.probability.hit;
+
+/**
+ * Input Boundary for actions which are related to collecting win probabilities when hitting.
+ */
+public interface ProbabilityHitInputBoundary {
+
+    /**
+     * Executes the probabilityHit usecase
+     */
+    void execute();
+}
