+++ src/main/java/use_case/probability/hit/ProbabilityHitOutputData.java
@@ -1,0 +1,14 @@
+package use_case.probability.hit;
+
+public class ProbabilityHitOutputData {
+
+    private final int hitWinProbability;
+
+    ProbabilityHitOutputData(int hitWinProbability) {
+        this.hitWinProbability = hitWinProbability;
+    }
+
+    public int getHitWinProbability() {
+        return hitWinProbability;
+    }
+}
