+++ src/main/java/use_case/probability/stand/ProbabilityStandInputBoundary.java
@@ -1,0 +1,12 @@
+package use_case.probability.stand;
+
+/**
+ * Input Boundary for actions which are related to collecting win probabilities when standing.
+ */
+public interface ProbabilityStandInputBoundary {
+
+    /**
+     * Executes the probabilityStand usecase.
+     */
+    void execute();
+}
