+++ src/main/java/use_case/probability/stand/ProbabilityStandOutputData.java
@@ -1,0 +1,14 @@
+package use_case.probability.stand;
+
+public class ProbabilityStandOutputData {
+
+    private final int standWinProbability;
+
+    public ProbabilityStandOutputData(int standWinProbability) {
+        this.standWinProbability = standWinProbability;
+    }
+
+    public int getStandWinProbability() {
+        return standWinProbability;
+    }
+}
