+++ src/main/java/view/ProbabilityStandView.java
@@ -1,11 +1,13 @@
 package view;
 
+import interface_adapter.probability.stand.ProbabilityStandViewModel;
+
 import java.awt.Font;
 import java.awt.BorderLayout;
 
 import javax.swing.*;
 
-import static interface_adapter.assisted_mode.AssistedModeColourConstants.TABLECOLOUR;
+import static interface_adapter.probability.ProbabilityColourConstants.TABLECOLOUR;
 
 /**
  * This is the view presenting the information of the ProbabilityStand Use Case.
@@ -13,13 +15,16 @@ import static interface_adapter.assisted_mode.AssistedModeColourConstants.TABLEC
  */
 public class ProbabilityStandView extends JPanel {
 
+    private final ProbabilityStandViewModel probabilityStandViewModel;
 
-    public ProbabilityStandView(int outputData) {
+    public ProbabilityStandView(ProbabilityStandViewModel probabilityStandViewModel) {
+        this.probabilityStandViewModel = probabilityStandViewModel;
 
         setLayout(new BorderLayout());
         setSize(300, 100);
         setBackground(TABLECOLOUR);
 
+        int outputData = probabilityStandViewModel.getState().getStandProbability();
 
         JLabel hitLabel = new JLabel("<html><font color = 'white'>Stand: " + outputData + "%</font></html>");
         hitLabel.setFont(new Font("Times New Roman", Font.BOLD, 14));
